+++ README.md
@@ -1,2 +1,2 @@
-# ec_dh_schnorr_lea_cbc
-Secure access to dataset of sound files. Password 2 level access to the data, encr/decr by LEA in CBC mode; secure channel for secret key generation with EC DH + Schnorr signature
+# lea_python
+
